pyproject.toml
[project]
name = "meshnetmap"
version = "1.0.0"
description = "Map and visualize Meshtastic network topology via Bluetooth"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "meshtastic>=2.3.0",
    "bleak>=0.21.0",
    "networkx>=3.0",
    "matplotlib>=3.7.0",
    "plotly>=5.18.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
]

[project.scripts]
meshnetmap = "meshnetmap:main"

[tool.setuptools]
packages = ["meshnetmap", "meshnetmap.collector", "meshnetmap.visualizer"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]